Huge milestone! Working awesome :)

+++ src/main.rs
@@ -5,40 +5,49 @@ mod cli;
 
 use std::sync::LazyLock;
 
-use cli::{CliArgs, HydePathArgs};
+use cli::CliArgs;
 use hyde_lib::{error::ResultExtensions, HydeConfig};
 use notify::Watcher;
 
-// pub static ARGS: LazyLock<CliArgs> = LazyLock::new(<CliArgs as clap::Parser>::parse);
+pub static ARGS: LazyLock<CliArgs> = LazyLock::new(<CliArgs as clap::Parser>::parse);
 
 fn main() {
 	use cli::HydeCommand::*;
 
-	let args = <CliArgs as clap::Parser>::parse();
-
-	match args.command {
-		Completion { shell } => cli::generate_completion(shell),
-		Build(paths) => hyde_lib::process_project(&init_config(paths)).handle_as_error(),
-		Serve(paths) => serve(init_config(paths)),
-		Clean(paths) => if let HydeConfig { output_dir, .. } = init_config(paths) && output_dir.exists() {
-			std::fs::remove_dir_all(&output_dir).handle_as_error();
+	match &ARGS.command {
+		Completion { shell } => cli::generate_completion(*shell),
+		Build => hyde_lib::process_project(&init_config()).handle_as_error(),
+		Serve { port } => serve(init_config(), port.unwrap_or(4000)),
+		Clean => {
+			let HydeConfig { output_dir, .. } = init_config();
+			if output_dir.exists() {
+				std::fs::remove_dir_all(&output_dir).handle_as_error();
+			}
 		}
 	}
 }
 
-fn init_config(paths: HydePathArgs) -> HydeConfig {
-	let project_dir = paths.dir.map_or(std::env::current_dir().unwrap(), |p| std::env::current_dir().unwrap().join(p));
+fn init_config() -> HydeConfig {
+	let paths = &ARGS.path_args;
+
+	let project_dir = paths.dir.as_ref().map_or(std::env::current_dir().unwrap(), |p| std::env::current_dir().unwrap().join(p));
+
+	// Check for the 'hyde.yml' file.
+	if !project_dir.join("hyde.yml").exists() {
+		eprintln!("No 'hyde.yml' file found in the project directory.");
+		std::process::exit(1);
+	}
 
 	hyde_lib::HydeConfig {
 		source_dir: project_dir.join("src"),
-		output_dir: paths.out.unwrap_or(project_dir.join("site")),
+		output_dir: paths.out.to_owned().unwrap_or(project_dir.join("site")),
 		plugins_dir: project_dir.join("plugins"),
 		layout_dir: project_dir.join("layouts"),
 		project_dir,
 	}
 }
 
-fn serve(config: HydeConfig) {
+fn serve(config: HydeConfig, port: u16) {
 	let (tx, rx) = std::sync::mpsc::channel::<notify::Result<notify::Event>>();
 	let mut watcher = notify::recommended_watcher(tx).unwrap();
 	watcher.watch(&config.source_dir, notify::RecursiveMode::Recursive).unwrap();
@@ -56,11 +65,11 @@ fn serve(config: HydeConfig) {
 	let serve_path = config.output_dir.clone();
 	let server = async move || {
 		let app = axum::Router::new()
-			.nest_service("/", tower_http::services::ServeDir::new(serve_path))
+			.fallback_service(tower_http::services::ServeDir::new(serve_path))
 			.layer(reload_layer)
 			;
 
-		let listener = tokio::net::TcpListener::bind("0.0.0.0:4000").await.unwrap();
+		let listener = tokio::net::TcpListener::bind(("0.0.0.0", port)).await.unwrap();
 		axum::serve(listener, app).await.unwrap();
 	};
 	
@@ -80,9 +89,7 @@ fn serve(config: HydeConfig) {
 		}
 
 		if to_reload {
-			if let Err(e) = hyde_lib::process_project(&config) {
-				println!("{e}");
-			}
+			hyde_lib::process_project(&config).handle_as_error();
 
 			reload_handle.reload();
 			to_reload = false;
